notes/pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.14"
fastapi = "^0.124.4"
uvicorn = {extras = ["standart"], version = "^0.38.0"}
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "^2.12.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.45"}
asyncpg = "^0.31.0"
alembic = "^1.17.2"
orjson = "^3.11.5"
celery = "^5.6.2"
kombu = "^5.6.2"
sqlalchemy-utc = "^0.14.0"
aiobotocore = "^3.1.0"
python-multipart = "^0.0.21"
fastapi-debug-toolbar = "^0.6.3"

[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)"
]
